ai_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
transformers==4.57.5
torch==2.7.1
pydantic==2.9.2
python-multipart==0.0.17

# Core ML / Math
numpy==2.2.6
pillow==11.1.0
scipy==1.15.2

# Hugging Face / NLP
tokenizers==0.22.2
huggingface-hub==0.36.0
safetensors==0.5.3

# Utilities
requests==2.33.0
tqdm==4.67.1
python-dotenv==1.0.1

# FastAPI dependencies
starlette==0.37.2
typing_extensions==4.12.2
anyio==4.12.1
scikit-learn==1.6.1
easyocr

pillow-avif-plugin==1.4.3